+++ kadai1/lib/mycalc.go
@@ -1,0 +1,11 @@
+package lib
+
+import "fmt"
+
+func Rectangle(l int, m int) int {
+	return l * m
+}
+
+func Intro() {
+	fmt.Println("これはmypkgです。")
+}
